pyproject.toml
[project]
name = "telegram-chatbot-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "fastapi[standard]>=0.115.13",
    "python-dotenv>=1.1.0",
]
